created the study_searches table in the ctgov schema and moved the searches functionality over to it
