--- db/migrate/20200922181240_create_searches.rb
@@ -1,12 +1,0 @@
-class CreateSearches < ActiveRecord::Migration
-  def change
-    create_table 'support.searches' do |t|
-      t.boolean :save_tsv, null: false, default: false
-      t.string :query, null: false
-      t.string :grouping, null: false, default: ''
-
-      t.timestamps null: false
-    end
-    add_index :searches, [:query, :grouping], unique: true
-  end
-end
