--- db/migrate/20201201235004_add_name_to_searches.rb
@@ -1,9 +1,0 @@
-class AddNameToSearches < ActiveRecord::Migration
-  def up
-    add_column :searches, :name, :string, null: false, default: ''
-  end
-
-  def down
-    remove_column :searches, :name, :string
-  end
-end
