--- db/migrate/20201207173334_add_beta_api_to_searches.rb
@@ -1,9 +1,0 @@
-class AddBetaApiToSearches < ActiveRecord::Migration
-  def up
-    add_column :searches, :beta_api, :boolean, default: false, null: false
-  end
-
-  def down 
-    remove_column :searches, :beta_api
-  end
-end
