--- db/migrate/20210108171515_add_search_id_to_categories.rb
@@ -1,8 +1,0 @@
-class AddSearchIdToCategories < ActiveRecord::Migration
-  def up
-    add_column :categories, :search_id, :integer, foreign_key: true
-  end
-  def down
-    remove_column :categories, :search_id
-  end
-end
