+++ db/migrate/20210108195415_create_study_searches.rb
@@ -1,0 +1,14 @@
+class CreateStudySearches < ActiveRecord::Migration
+  def change
+    create_table :study_searches do |t|
+      t.boolean :save_tsv, null: false, default: false
+      t.string :query, null: false
+      t.string :grouping, null: false, default: ''
+      t.string :name, null: false, default: ''
+      t.boolean :beta_api, null: false, default: false
+      
+      t.timestamps null: false
+    end
+    add_index :study_searches, [:query, :grouping], unique: true
+  end
+end
