+++ db/migrate/20210108200600_add_study_search_id_to_categories.rb
@@ -1,0 +1,5 @@
+class AddStudySearchIdToCategories < ActiveRecord::Migration
+  def change
+    add_column :categories, :study_search_id, :integer, foreign_key: true
+  end
+end
