+++ spec/factories.rb
@@ -1,4 +1,7 @@
 FactoryGirl.define do
+  factory :study_search do
+    
+  end
   factory :study_json_record do
     study_batch ""
     studies_saved_at "2019-11-25 15:52:10"
