+++ spec/models/study_search_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Search, type: :model do
+RSpec.describe StudySearch, type: :model do
   pending "add some examples to (or delete) #{__FILE__}"
 end
